Add target determination to slang util.
